bookstore/requirement.txt
Django==2.0.5
djangorestframework==3.8.2
pkg-resources==0.0.0
Pygments==2.2.0
PyMySQL==0.8.0
pytz==2018.4
